# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Sura Codex API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /essays:
    get:
      operationId: listEssays
      tags: [essays]
      summary: List all essays
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of essays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Essay"
    post:
      operationId: createEssay
      tags: [essays]
      summary: Create an essay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EssayInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Essay"
  /essays/{id}:
    get:
      operationId: getEssay
      tags: [essays]
      summary: Get an essay by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Essay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Essay"
        "404":
          description: Not found
    patch:
      operationId: updateEssay
      tags: [essays]
      summary: Update an essay
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EssayUpdate"
      responses:
        "200":
          description: Updated essay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Essay"
    delete:
      operationId: deleteEssay
      tags: [essays]
      summary: Delete an essay
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /novels:
    get:
      operationId: listNovels
      tags: [novels]
      summary: List all novels
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of novels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Novel"
    post:
      operationId: createNovel
      tags: [novels]
      summary: Create a novel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NovelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Novel"
  /novels/{id}:
    get:
      operationId: getNovel
      tags: [novels]
      summary: Get a novel by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Novel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Novel"
        "404":
          description: Not found
    patch:
      operationId: updateNovel
      tags: [novels]
      summary: Update a novel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NovelUpdate"
      responses:
        "200":
          description: Updated novel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Novel"
    delete:
      operationId: deleteNovel
      tags: [novels]
      summary: Delete a novel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a piece of content
      parameters:
        - name: essayId
          in: query
          schema:
            type: integer
        - name: novelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (admin or own)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /comments/{id}/approve:
    patch:
      operationId: approveComment
      tags: [comments]
      summary: Admin approve a pending comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /likes/toggle:
    post:
      operationId: toggleLike
      tags: [likes]
      summary: Toggle a like on an essay or novel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LikeToggleInput"
      responses:
        "200":
          description: Like state after toggle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeToggleResult"
  /likes/counts:
    get:
      operationId: getLikeCounts
      tags: [likes]
      summary: Get like counts for a content item
      parameters:
        - name: essayId
          in: query
          schema:
            type: integer
        - name: novelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Like counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeCounts"
...
components:
...
  schemas:
...
    Essay:
      type: object
      required: [id, title, slug, excerpt, content, category, publishedAt, featured, readingTime, likeCount, commentCount, shareCount]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        coverImage:
          type: ["string", "null"]
        publishedAt:
          type: string
        featured:
          type: boolean
        readingTime:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        shareCount:
          type: integer
    EssayInput:
      type: object
      required: [title, excerpt, content, category]
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        coverImage:
          type: string
        featured:
          type: boolean
        readingTime:
          type: integer
    EssayUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        coverImage:
          type: string
        featured:
          type: boolean
        readingTime:
          type: integer
    Novel:
      type: object
      required: [id, title, slug, synopsis, content, publishedAt, featured, status, likeCount, commentCount, shareCount]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        synopsis:
          type: string
        content:
          type: string
        coverImage:
          type: ["string", "null"]
        publishedAt:
          type: string
        featured:
          type: boolean
        status:
          type: string
        chaptersCount:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        shareCount:
          type: integer
    NovelInput:
      type: object
      required: [title, synopsis, content, status]
      properties:
        title:
          type: string
        slug:
          type: string
        synopsis:
          type: string
        content:
          type: string
        coverImage:
          type: string
        featured:
          type: boolean
        status:
          type: string
        chaptersCount:
          type: integer
    NovelUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        synopsis:
          type: string
        content:
          type: string
        coverImage:
          type: string
        featured:
          type: boolean
        status:
          type: string
        chaptersCount:
          type: integer
...
    Comment:
      type: object
      required: [id, content, authorName, createdAt, isApproved]
      properties:
        id:
          type: integer
        essayId:
          type: ["integer", "null"]
        novelId:
          type: ["integer", "null"]
        parentId:
          type: ["integer", "null"]
        userId:
          type: ["string", "null"]
        authorName:
          type: string
        authorImage:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
        isApproved:
          type: boolean
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    CommentInput:
      type: object
      required: [content, authorName]
      properties:
        essayId:
          type: integer
        novelId:
          type: integer
        parentId:
          type: integer
        authorName:
          type: string
        content:
          type: string
    LikeToggleInput:
      type: object
      properties:
        essayId:
          type: integer
        novelId:
          type: integer
    LikeToggleResult:
      type: object
      required: [liked, count]
      properties:
        liked:
          type: boolean
        count:
          type: integer
    LikeCounts:
      type: object
      required: [count, userLiked]
      properties:
        count:
          type: integer
        userLiked:
          type: boolean
...