# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Sura Codex API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  /** @nullable */
  essayId?: number | null;
  /** @nullable */
  novelId?: number | null;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  userId?: string | null;
  authorName: string;
  /** @nullable */
  authorImage?: string | null;
  content: string;
  createdAt: string;
  isApproved: boolean;
  replies?: Comment[];
}