# lib/api-zod/src/generated/types/essay.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Sura Codex API
 * OpenAPI spec version: 0.1.0
 */

export interface Essay {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  category: string;
  /** @nullable */
  coverImage?: string | null;
  publishedAt: string;
  featured: boolean;
  readingTime: number;
  likeCount: number;
  commentCount: number;
  shareCount: number;
}