# lib/api-zod/src/generated/types/likeToggleResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Sura Codex API
 * OpenAPI spec version: 0.1.0
 */

export interface LikeToggleResult {
  liked: boolean;
  count: number;
}